Add workspace move subcommand (#96)
## Why is this change being made?

This new subcommand intends to support scenarios where the index is
built on a build server (such as D:\os\src) and then consumed on a
developer machine with a different path (such as C:\rs\src or
C:\repos\os\src), by allowing the workspace to be moved to the correct
local path for that user after it's been generated.

## What changed?

- Added a new endpoint (`workspace/move`) which calls `workspace.Move`
to set the workspace location.
- Added a new subcommand, `haystack workspace move <id> <new path>`,
that uses the new endpoint.

## How was the change tested?

Running locally:
<img width="1275" height="772" alt="image"
src="https://github.com/user-attachments/assets/c6a3cd4b-bae9-41ba-913d-99d62a44f3fc"
/>

A later sync-all has the expected behavior: timestamps have changed, but
the content hash has not, so indexing does not re-occur.