Standardize -h and --help experience across CLI (#102)
This PR standardizes the `-h` and `--help` flags experience.

If we were using Rust for this application, we'd be using
[`clap`](https://github.com/clap-rs/clap) and would get this kind of
consistency out of the box. Perhaps there is a Go CLI library we could
be using?

In the meantime, I was struggling initially with not having the CLI
respect standard help usage. This should make it easier to get started
using the CLI.

---------

Co-authored-by: Kyle Rader <kyrader@microsoft.com>